+++ huntflow_api_client/models/consts.py
@@ -13,3 +13,13 @@ class MemberType(str, Enum):
     owner = "owner"
     manager = "manager"
     watcher = "watcher"
+
+
+class VacancyState(str, Enum):
+    OPEN = "OPEN"
+    CLOSED = "CLOSED"
+    HOLD = "HOLD"
+    REOPEN = "REOPEN"
+    VACANCY_REQUEST_ATTACH = "VACANCY_REQUEST_ATTACH"
+    RESUME = "RESUME"
+    CREATED = "CREATED"
