+++ huntflow_api_client/models/response/vacancy_requests.py
@@ -7,7 +7,7 @@ from enum import Enum
 from pydantic import BaseModel, EmailStr, Field, PositiveInt
 
 from huntflow_api_client.models.common import PaginatedResponse
-from huntflow_api_client.models.utils.fields import File
+from huntflow_api_client.models.fields import File
 
 
 class VacancyRequestStatus(str, Enum):
